- Fixes issues with Department Code, DTR, Payslip - Adds Backup function - Adds Salary Grade Steps to Add and edit Function of Employee

+++ database/migrations/2023_05_11_522746_create_salary_grade_steps_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('salary_grade_id')->constrained('sgrades')->cascadeOnDelete();
             $table->string('step');
-            $table->decimal('amount',);
+            $table->double('amount',);
             $table->timestamps();
         });
     }
